// src/user/entities/user-state.enum.ts
export enum UserState {
  REGISTRATION_INITIAL = 'REGISTRATION_INITIAL',
  REGISTRATION_COMPLETE = 'REGISTRATION_COMPLETE',
  DATA_PRIVACY_CONFIRMATION = 'DATA_PRIVACY_CONFIRMATION',
  NAME = 'NAME',
  NAME_CONFIRMATION = 'NAME_CONFIRMATION',
  PHONE_NUMBER = 'PHONE_NUMBER',
  PHONE_NUMBER_CONFIRMATION = 'PHONE_NUMBER_CONFIRMATION',
  EMAIL = 'EMAIL',
  EMAIL_CONFIRMATION = 'EMAIL_CONFIRMATION',
  TAXPAYER_NUMBER = 'TAXPAYER_NUMBER',
  TAXPAYER_NUMBER_CONFIRMATION = 'TAXPAYER_NUMBER_CONFIRMATION',
  ADDRESS = 'ADDRESS',
  ADDRESS_CONFIRMATION = 'ADDRESS_CONFIRMATION',
}